pyproject.toml
[tool.poetry]
name = "ansible-waldur-generator"
version = "0.0.1"
description = "Ansible Waldur module generator."
authors = ["OpenNode Team <info@opennodecloud.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://waldur.com"
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.scripts]
ansible-waldur-generator = "ansible_waldur_generator.cli:main"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pyyaml = "^6.0"
jinja2 = "^3.1.2"
click = "^8.1.3"
waldur-api-client = "^7.7.4"
ansible = "^11.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins.ansible_waldur_generator]
"crud" = "ansible_waldur_generator.plugins.crud.plugin:CrudPlugin"
"facts" = "ansible_waldur_generator.plugins.facts.plugin:FactsPlugin"
